requirements.txt
Pillow==10.0.0
requests==2.31.0
python-bidi==0.4.2
arabic-reshaper==3.0.0
